pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "azure-foundry-architect-framework"
version = "1.0.0"
description = "Multi-agent framework for cloud architecture analysis, migration planning, and optimization using Azure AI Foundry"
requires-python = ">=3.11"

dependencies = [
    # Azure AI
    "azure-ai-projects>=1.0.0b10",
    "azure-ai-agents>=1.0.0b4",
    "azure-identity>=1.17.0",
    "azure-search-documents>=11.6.0",
    "azure-cosmos>=4.7.0",
    "azure-keyvault-secrets>=4.8.0",
    "azure-monitor-opentelemetry>=1.6.0",
    # Semantic Kernel
    "semantic-kernel>=1.16.0",
    # OpenAI
    "openai>=1.55.0",
    # API
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "python-multipart>=0.0.12",
    # Tools
    "httpx>=0.27.0",
    "aiofiles>=24.1.0",
    "python-dotenv>=1.0.1",
    "PyYAML>=6.0.2",
    "rich>=13.9.0",
    "structlog>=24.4.0",
    # Code Analysis
    "tree-sitter>=0.23.0",
    "gitpython>=3.1.43",
    # IaC Parsing
    "hcl2>=4.3.4",
    "jsonschema>=4.23.0",
    # Reporting
    "jinja2>=3.1.4",
    "weasyprint>=62.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-httpx>=0.32.0",
    "ruff>=0.7.0",
    "mypy>=1.13.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
asyncio_mode = "auto"
